[Travis] Do not use --quiet when downloading
I think it used to be the case that Travis would output one line per % of the
progression, which filled the console with crap, but now it seems to only show
a line that is updated in place. Hence, there's no need to be quiet anymore
when downloading, and that makes it easier to debug failures.